pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cloud_workflows_app"
repository = "https://github.com/samarth1029/authentiChatAI.git"
version = "0.1.0"
description = "Python+FastApi based app for product authentication catbot"
authors = ["Pramit Dash <pramit.dash@stryze.com>"]
license = "Proprietary"
readme = "README.md"
keywords = ["authenti_chatai", "cloudbuild"]
packages = [
    { include = "app" },
    { include = "app/**/*.py" },
]

[[tool.poetry.source]]
name = "default_pypi"
url = "https://pypi.python.org/simple"
secondary = false

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pandas = ">=1.3.2"
python-dateutil = "2.8.2"
requests = "2.28.1"
pytz = ">=2021.1,<2022.0"
pydantic = ">=1.9.1,<1.10.0"
python-dotenv = ">=0.20.0,<0.21.0"
numpy = ">=1.23.0,<1.24.0"
tabulate = ">=0.8.9,<0.9.0"
fastapi = ">=0.79.0,<0.80.0"
uvicorn = ">=0.18.2,<0.19.0"
pytest = ">=7.1.2"
starlette = "0.19.1"
db-dtypes = "~1.0.4"
google-api-core = ">=2.8.2,<2.9.0"
google-cloud-bigquery = ">=3.3.0,<3.4.0"
google-cloud-secret-manager = ">=2.10.0,<2.11.0"
google-cloud-logging = "^3.2.5"
google-cloud-storage = "^2.5.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = ">=7.1"
cookiecutter = "^1.7.2"
tox = "^3.23.0"
bump2version = "^1.0.1"
flake8 = "^3.9.1"
pip = "^21.1"
wheel = "^0.36.2"
coverage = {extras = ["toml"], version = "^5.5"}
atomicwrites = "^1.4.1"


[tool.poetry.scripts]
authenti_chatai= "app.__main__:main"

